+++ devices/switch/on_off_switch.py
@@ -12,3 +12,13 @@ class OnOffSwitch(Device):
     def get_string_value(self, value, device):
         return 'On' if value.lower() == 'on' else 'Off'
         
+    def handle_command(self, device_data, command, level, color):
+        device_address = device_data['ieee_addr']
+        device = self.get_device(device_address, self.alias)
+
+        Domoticz.Debug('Command "' + command + '" from device "' + device.Name + '"')
+
+        device.Update(
+            nValue=self.get_numeric_value(command, device),
+            sValue=self.get_string_value(command, device)
+        )